Move the .ts files in src into a common subdirectory.  The common.ts file is renames to common/views.ts as that is more appropriate for what it is.
