+++ src/components/common/Problem.vue
@@ -55,7 +55,7 @@
 import { defineComponent, ref, watch, onMounted, nextTick, PropType } from 'vue';
 import type { SubmitButton } from 'src/typings/renderer';
 import { fetchProblem, RendererParams } from 'src/api-requests/renderer';
-import { RENDER_URL } from 'src/constants';
+import { RENDER_URL } from 'src/common/constants';
 import * as bootstrap from 'bootstrap';
 import type JQueryStatic from 'jquery';
 import JQuery from 'jquery';
