+++ src/pages/instructor/ProblemSetsManager.vue
@@ -48,7 +48,7 @@
 import { defineComponent, computed, ref } from 'vue';
 import { useStore } from 'src/store';
 import { ProblemSet, QuizDates } from 'src/store/models/problem_sets';
-import { formatDate } from 'src/common';
+import { formatDate } from 'src/common/views';
 
 export default defineComponent({
 	name: 'ProblemSetsManager',
